Fixing errors again on TrainServer

+++ src/main/java/mtr/data/TrainServer.java
@@ -45,7 +45,7 @@ public class TrainServer extends Train {
 	}
 
 	@Override
-	protected void startUp(World world, int trainLength, int trainSpacing, boolean isOppositeRail) {
+	protected void startUp(World world, int trainLength, float trainSpacing, boolean isOppositeRail) {
 		canDeploy = false;
 		isOnRoute = true;
 		stopCounter = 0;
@@ -54,7 +54,7 @@ public class TrainServer extends Train {
 			railProgress += trainLength * trainSpacing;
 			reversed = !reversed;
 		}
-		nextStoppingIndex = getNextStoppingIndex(trainSpacing);
+		nextStoppingIndex = getNextStoppingIndex((int)trainSpacing);
 	}
 
 	@Override
@@ -208,7 +208,7 @@ public class TrainServer extends Train {
 		if (update) {
 			final int offsetTicks = isOnRoute ? 0 : depot.getNextDepartureTicks(Depot.getHour(world));
 			if (offsetTicks >= 0) {
-				writeArrivalTimes(writeScheduleCallback, isUnlimited, depot.routeIds, offsetTicks, trainMapping, trainLength, trainMapping.trainType.getSpacing());
+				writeArrivalTimes(writeScheduleCallback, isUnlimited, depot.routeIds, offsetTicks, trainMapping, trainLength, (int) trainMapping.trainType.getSpacing());
 			}
 		}
 
@@ -217,7 +217,7 @@ public class TrainServer extends Train {
 
 	public void writeTrainPositions(Set<UUID> trainPositions) {
 		if (!path.isEmpty()) {
-			final int trainSpacing = trainMapping.trainType.getSpacing();
+			final float trainSpacing = trainMapping.trainType.getSpacing();
 			final int headIndex = getIndex(0, trainSpacing, true);
 			final int tailIndex = getIndex(trainLength, trainSpacing, false);
 			for (int i = tailIndex; i <= headIndex; i++) {
